requirements.txt
appnope==0.1.3
asttokens==2.4.1
bcrypt==3.1.7
certifi==2020.12.5
cffi==1.16.0
click==7.1.2
comm==0.2.0
crewai
crewai-tools
cryptography==41.0.7
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.7
executing==2.0.1
filelock==3.13.1
Flask==1.1.2
ifaddr==0.1.7
ipykernel==6.26.0
ipython==8.17.2
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==2.11.2
jupyter_client==8.6.0
jupyter_core==5.5.0
jsonpath-ng==1.6.1
jsonschema==4.23.0
MarkupSafe==1.1.1
matplotlib-inline==0.1.6
mendel-development-tool==1.5.2
nest-asyncio==1.5.8
packaging==23.2
paramiko==2.7.1
parso==0.8.3
pexpect==4.8.0
platformdirs==3.11.0
prompt-toolkit==3.0.41
psutil==5.9.6
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.16.1
PyNaCl==1.4.0
PyYAML
python-dateutil==2.8.2
python-dotenv
pyzmq==25.1.1
six==1.16.0
stack-data==0.6.3
tornado==6.3.3
traitlets==5.13.0
virtualenv==20.24.6
wcwidth==0.2.10
Werkzeug==1.0.1
zeroconf==0.28.0
